% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generators.R
\name{Anything}
\alias{Anything}
\title{Convenience wrapper around many different generators}
\usage{
Anything()
}
\description{
A wrapper around
\preformatted{
(Bool() \%>\% withNA) \%|\%
(String() \%>\% withNA) \%|\%
(SingleCharacter() \%>\% withNA) \%|\%
(NumericInRange(-1e6,1e6) \%>\% withZero \%>\% withNA) \%|\%
(IntegerInRange(-1e6,1e6) \%>\% withZero \%>\% withNA)
}
}

