% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generators.R
\name{AsOne}
\alias{AsOne}
\title{Wrapper for vector generators to be used 'as one' insted of element-wise}
\usage{
AsOne(x)
}
\description{
To be used if the generated vector should be used as whole vector rather
than its elements one-by-one.
}

