% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generators.R
\name{DataFrame}
\alias{DataFrame}
\title{Generator for data frames}
\usage{
DataFrame(...)
}
\description{
It may contain other generators.
The elements of the data frame must
have the equal number of elements
(equal the number of rows of the data frame).
}

