% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/docgen.R
\name{docgen}
\alias{docgen}
\title{Automatically generate package documentation from a gentest() result}
\usage{
docgen(gentest_result, title = attr(gentest_result, "function"))
}
\arguments{
\item{title}{String, the title of the .
Default: a function name stored in \code{gentest_result}.

The documentation file is saved as \code{XXX.Rd} -- where
\code{XXX} is determined by the function name stored in
the \code{gentest_result} -- in the sub-directory \code{man}
of the current working directory. This sub-directory needs to be
created first, if it does not yet exist.}

\item{gentest_restult}{Value returned by \code{\link[gentest]{gentest}}.}
}
\description{
Generates automatically the documentation for a function tested with
\code{\link[gentest]{gentest}}. Takes into consideration only those
results which did not return errors during tests.
of the current working directory
}

