% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core.R
\name{gentest}
\alias{gentest}
\title{Test a function with different combinations of arguments, possibly generated randomly}
\usage{
gentest(..f, ..p = identity, ...)
}
\arguments{
\item{..f}{A non-anonymous function to be tested
(and possibly to be later documented with \code{\link[gentest]{docgen}}).}

\item{..p}{An optional function with one argument (i.e. arity 1)
to be applied to the value retunred by function \code{f}. By
default \code{p = \link{identity}}. Useful for applying additional
tests/checks that do not have to or should not be included in function \code{f}.}

\item{...}{Different values to be applied as arguments
under different calls of function \code{f}. If vectors are to be used
as single objects (rather than single value by value) they should be wrapped in
a \code{\link{list}}, or \code{\link[gentest]{AsOne}}
(an alias for \code{list} for clarity), or \code{\link[gentest]{Replicate}}.}
}
\value{
A \code{data.frame} (S3 class \code{gentest_result}) for all
the combinations of argument values specified
in \code{...}, including:
\itemize{
\item the columns with the argument values
\item return values (column \code{..value}),
\item warnings (columns \code{..warning}),
\item errors (columns \code{..error}),
\item the character columns with the suffix \code{..type}
which describe the class
}
}
\description{
Test a function with different combinations of arguments, possibly generated randomly
}

