% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generators.R
\name{IntegerInRange}
\alias{IntegerInRange}
\title{Generator for integer numbers based on sample()}
\usage{
IntegerInRange(bottom, top, vector_length = 100)
}
\arguments{
\item{bottom}{minimum values that may be generated}

\item{top}{maximum values that may be generated}

\item{vector_length}{A length of the vector to be generated.
It must be >= 1.}
}
\description{
Generator for integer numbers based on sample()
}

