% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generators.R
\name{Name}
\alias{Name}
\title{Add a name, to be used ulitmately by docgen(), to the value(s) used in the gentest()}
\usage{
Name(obj, name)
}
\arguments{
\item{obj}{A return value of a generator function}

\item{name}{String, a human readable name for the generator}
}
\description{
Add a name, to be used ulitmately by docgen(), to the value(s) used in the gentest()
}

