% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generators.R
\name{NumericInRange}
\alias{NumericInRange}
\title{Generator for numbers (doubles) based on runif()}
\usage{
NumericInRange(bottom, top, vector_length = 100)
}
\arguments{
\item{bottom}{minimum values that may be generated}

\item{top}{maximum values that may be generated}

\item{vector_length}{A length of the vector to be generated.
It must be >= 1.}
}
\description{
Generator for numbers (doubles) based on runif()
}

