% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generators.R
\name{Replicate}
\alias{Replicate}
\title{Wrapper for vector generators replicating randomly whole vectors}
\usage{
Replicate(expr, times = 10)
}
\description{
To be used to generate different vectors of the same type/class.
}

