% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generators.R
\name{Scalar}
\alias{Scalar}
\title{Modifier reducing vector to scalar (a single element)}
\usage{
Scalar(vec)
}
\description{
Cuts the length of the generated vector \code{vec} to 1
by selecting a random element of the vector.
}

