% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generators.R
\name{SingleCharacter}
\alias{SingleCharacter}
\title{Generator for single characters}
\usage{
SingleCharacter(vector_length = 100)
}
\arguments{
\item{vector_length}{A length of the vector to be generated.
It must be >= 1.
The generated characters are only from the ASCII code range 32--126.}
}
\description{
Generator for single characters
}

