% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core.R
\name{stopIfErrOrWarn}
\alias{stopIfErrOrWarn}
\title{Convenience wrapper around stopIfErr() and stopIfWarn()}
\usage{
stopIfErrOrWarn(x)
}
\description{
Simply \preformatted{
library(magrittr)
x \%>\%
    stopIfErr \%>\%
    stopIfWarn
}
}

