% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generators.R
\name{String}
\alias{String}
\title{Generator for strings}
\usage{
String(string_length = replicate(vector_length, sample.int(101, 1) - 1),
  vector_length = 100)
}
\arguments{
\item{string_length}{A numeric vector of string lengths
(in terms of the numbers of characters) in a vector
to be generated. Each element must be >= 0.}

\item{vector_length}{A length of the vector to be generated.
It must be >= 1.

length of \code{string_length} must be equal \code{vector_length}
or \code{string_length} must be a scalar (\code{length(string_length) == 1}).
The generated strings include only ASCII characters in the code range 32--126.}
}
\description{
Generator for strings
}

