% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generators.R
\name{withNA}
\alias{withNA}
\title{Modifier injecting missing observations (NAs)}
\usage{
withNA(vec, share = 0.1)
}
\description{
Injects randomly missing observations (NA values)
replacing about 100*\code{share} per cent of
elements.
}

