% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generators.R
\name{withZero}
\alias{withZero}
\title{Modifier injecting observations with 0 (zeros)}
\usage{
withZero(vec, share = 0.1)
}
\description{
Injects randomly 0 (zeros)
replacing about 100*\code{share} per cent of
elements.
}

